Auto aim first testing, need to fix the drop down
