--- Command_Library/src/main/java/Deposit.java
@@ -1,66 +1,0 @@
-import dev.weaponboy.command_library.CommandLibrary.Commands.LambdaCommand;
-import dev.weaponboy.command_library.CommandLibrary.Subsystem.SubSystem;
-import com.qualcomm.robotcore.hardware.DcMotor;
-import com.qualcomm.robotcore.hardware.DcMotorEx;
-import com.qualcomm.robotcore.hardware.Servo;
-
-import dev.weaponboy.command_library.CommandLibrary.Commands.LambdaCommand;
-import dev.weaponboy.command_library.CommandLibrary.Subsystem.SubSystem;
-
-
-public class Deposit extends SubSystem {
-    DcMotor deposite1;
-    Servo depo1;
-
-    @Override
-    public void init() {
-
-    }
-
-    @Override
-    public void execute() {
-
-    }
-
-
-//public LambdaCommand deposithigh = new LambdaCommand(
-//            () ->
-//
-//            () ->
-//
-//            () ->
-//
-//    );
-//  //  public LambdaCommand depositeLow = new LambdaCommand(
-//
-//            () ->
-//
-//
-//            () ->
-//
-//            () ->
-//
-//    );
-//
-//
-//    public LambdaCommand stow = new LambdaCommand(    () ->
-
-//            () ->
-//
-//            () ->
-//
-//
-//    );
-//
-}
-
-
-
-
-
-
-
-
-
-
-
