+++ Command_Library/src/main/java/dev/weaponboy/command_library/CommandLibrary/OpmodeEX/OpModeEX.java
@@ -8,18 +8,17 @@ import com.qualcomm.robotcore.util.ElapsedTime;
 import java.util.List;
 
 import dev.weaponboy.command_library.CommandLibrary.Subsystem.SubSystem;
-import dev.weaponboy.command_library.Subsystems.Collection2;
+import dev.weaponboy.command_library.Subsystems.Collection;
 import dev.weaponboy.command_library.Subsystems.Delivery;
 import dev.weaponboy.command_library.Subsystems.DriveBase;
 import dev.weaponboy.command_library.Subsystems.Hang;
 import dev.weaponboy.command_library.Subsystems.Odometry;
-import dev.weaponboy.command_library.Subsystems.Collection;
 
 public abstract class OpModeEX extends OpMode {
 
     public DriveBase driveBase = new DriveBase(this);
 
-    public Collection2 collection = new Collection2(this);
+    public Collection collection = new Collection(this);
 
     public Delivery delivery = new Delivery(this);
