+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Testing/TestingRotate.java
@@ -4,9 +4,8 @@ import com.qualcomm.robotcore.eventloop.opmode.OpMode;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 import com.qualcomm.robotcore.hardware.PwmControl;
 
-import dev.weaponboy.command_library.CommandLibrary.OpmodeEX.OpModeEX;
 import dev.weaponboy.command_library.Hardware.ServoDegrees;
-import dev.weaponboy.command_library.Subsystems.Collection2;
+
 @TeleOp
 public class TestingRotate extends OpMode {
